move all tests into separate folder at root level and update import paths to stop after foundations directory to ensure index files are covered by tests
