+++ test/foundations/typography.test.tsx
@@ -1,8 +1,12 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import { renderAndTestProps } from '../../../utilities/testing-utils';
-import { TEXT_ALIGN, TEXT_TRANSFORM, Fonts } from '..';
-import { SHIRT_SIZE } from '../../layout/box';
+import { renderAndTestProps } from '../../src/utilities/testing-utils';
+import {
+  SHIRT_SIZE,
+  Fonts,
+  TEXT_ALIGN,
+  TEXT_TRANSFORM,
+} from '../../src/foundations';
 
 describe('Typography', () => {
   it('Renders Default', () => {
